refactor: merge smoke-auth into unified smoke test suite
Move authenticated smoke tests from e2e/smoke-auth/ into
e2e/smoke/authenticated/ and collapse the separate Playwright
projects (chromium + smoke-auth) into a single smoke project
with setup dependency. One command runs all smoke tests.
