+++ apps/web/e2e/smoke/authenticated/admin-dev-tools.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../fixtures/authenticated';
+import { test, expect } from '../../fixtures/authenticated';
 
 test.use({ userRole: 'admin' });
