+++ apps/web/package.json
@@ -11,8 +11,7 @@
         "test": "vitest",
         "test:run": "vitest run",
         "test:e2e": "playwright test",
-        "test:e2e:smoke": "playwright test e2e/smoke",
-        "test:e2e:smoke-auth": "playwright test --project=setup --project=smoke-auth",
+        "test:e2e:smoke": "playwright test --project=setup --project=smoke",
         "test:e2e:admin": "playwright test --project=setup --project=admin",
         "test:e2e:ui": "playwright test --ui",
         "test:coverage": "vitest run --coverage",
